+++ claim_extractor/src/schemas/linked_trust.meta
@@ -1,0 +1,12 @@
+
+claim: one of 'impact', 'rated', 'same_as'
+aspect: similar to 'impact:financial', 'impact:social', 'impact:work', 'quality:overall', 'quality:affordable', 'quality:fun', 'report:scam', 'report:dangerous'
+howKnown: will be set by the caller later
+effectiveDate: when the impact or rating occurred, if you can tell
+object: only if it has a url
+subject: try to find a url (may not be possible), otherwise use text
+statement: this is the concise narrative of the claim, quoted from the text
+amt: optional, generally only for impact claims and only if an amount is present.  The total impact claimed.
+unit: include if amt is set ; usd if dollars, or set according to what amt is counting
+stars: optional if claim is a rating, scale is 1-5, otherwise leave out of the result
+score: if claim is a rating, may be set from -1 to 1 as a float, otherwise unset.  5 stars = score of 1
