+++ pdf_parser/requirements.txt
@@ -1,0 +1,140 @@
+aiohappyeyeballs==2.4.4
+aiohttp==3.11.8
+aiosignal==1.3.1
+annotated-types==0.7.0
+anthropic==0.39.0
+anyio==4.6.2.post1
+asgiref==3.8.1
+attrs==24.2.0
+backoff==2.2.1
+bcrypt==4.2.0
+build==1.2.2.post1
+cachetools==5.5.0
+certifi==2024.8.30
+cffi==1.17.1
+charset-normalizer==3.4.0
+chroma-hnswlib==0.7.6
+chromadb==0.5.18
+click==8.1.7
+coloredlogs==15.0.1
+cryptography==44.0.0
+dataclasses-json==0.6.7
+defusedxml==0.7.1
+Deprecated==1.2.14
+distro==1.9.0
+durationpy==0.9
+fastapi==0.115.5
+filelock==3.16.1
+flatbuffers==24.3.25
+frozenlist==1.5.0
+fsspec==2024.10.0
+google-auth==2.36.0
+googleapis-common-protos==1.66.0
+grpcio==1.67.1
+h11==0.14.0
+httpcore==1.0.6
+httptools==0.6.4
+httpx==0.27.2
+httpx-sse==0.4.0
+huggingface-hub==0.26.2
+humanfriendly==10.0
+idna==3.10
+importlib_metadata==8.5.0
+importlib_resources==6.4.5
+iniconfig==2.0.0
+Jinja2==3.1.4
+jiter==0.7.1
+joblib==1.4.2
+jsonpatch==1.33
+jsonpointer==3.0.0
+kubernetes==31.0.0
+langchain==0.3.9
+langchain-anthropic==0.3.0
+langchain-community==0.3.8
+langchain-core==0.3.21
+langchain-text-splitters==0.3.2
+langsmith==0.1.147
+linked-claims-extractor==0.1.1
+markdown-it-py==3.0.0
+MarkupSafe==3.0.2
+marshmallow==3.23.1
+mdurl==0.1.2
+mmh3==5.0.1
+monotonic==1.6
+mpmath==1.3.0
+multidict==6.1.0
+mypy-extensions==1.0.0
+networkx==3.4.2
+numpy==1.26.4
+oauthlib==3.2.2
+onnxruntime==1.20.0
+openai==1.55.3
+opentelemetry-api==1.28.1
+opentelemetry-exporter-otlp-proto-common==1.28.1
+opentelemetry-exporter-otlp-proto-grpc==1.28.1
+opentelemetry-instrumentation==0.49b1
+opentelemetry-instrumentation-asgi==0.49b1
+opentelemetry-instrumentation-fastapi==0.49b1
+opentelemetry-proto==1.28.1
+opentelemetry-sdk==1.28.1
+opentelemetry-semantic-conventions==0.49b1
+opentelemetry-util-http==0.49b1
+orjson==3.10.11
+overrides==7.7.0
+packaging==24.2
+pdfminer.six==20240706
+pillow==11.0.0
+pluggy==1.5.0
+posthog==3.7.0
+propcache==0.2.0
+protobuf==5.28.3
+pyasn1==0.6.1
+pyasn1_modules==0.4.1
+pycparser==2.22
+pydantic==2.9.2
+pydantic-settings==2.6.1
+pydantic_core==2.23.4
+Pygments==2.18.0
+PyMuPDF==1.24.13
+pypdf==5.1.0
+PyPika==0.48.9
+pyproject_hooks==1.2.0
+pytest==8.3.3
+python-dateutil==2.9.0.post0
+python-dotenv==1.0.1
+PyYAML==6.0.2
+rank-bm25==0.2.2
+regex==2024.11.6
+requests==2.32.3
+requests-oauthlib==2.0.0
+requests-toolbelt==1.0.0
+rich==13.9.4
+rsa==4.9
+safetensors==0.4.5
+scikit-learn==1.5.2
+scipy==1.14.1
+sentence-transformers==3.3.0
+shellingham==1.5.4
+six==1.16.0
+sniffio==1.3.1
+SQLAlchemy==2.0.35
+starlette==0.41.2
+sympy==1.13.1
+tenacity==9.0.0
+threadpoolctl==3.5.0
+tokenizers==0.20.3
+torch==2.5.1
+tqdm==4.67.0
+transformers==4.46.2
+typer==0.13.0
+typing-inspect==0.9.0
+typing_extensions==4.12.2
+urllib3==2.2.3
+uvicorn==0.32.0
+uvloop==0.21.0
+watchfiles==0.24.0
+websocket-client==1.8.0
+websockets==14.1
+wrapt==1.16.0
+yarl==1.18.0
+zipp==3.21.0
